Skip bad XML imports with warning.
Recently we were ingesting a very large set of marc xml records only to
have the whole import fail because the first record had a field with an
unknown value.

This change adds a new method `MARC::GenericPullParse#new_field` that is
used as a factory for creating new record fields in the xml parser. Then
when an error happens it is noted and the record is skipped instead of
breaking the entire import.

I stumbled onto this issue when working with the traject library.
I tried fixing it there, but could not figure it out without monkey
patching the marc library. So I think it makes more sense to update the
marc library.

Additional QA related change:

* Updates jruby version in .travis.yml to a specific version of jruby
available on Travis' Trusty image:

http://rubies.travis-ci.org/

See issue: https://github.com/travis-ci/travis-ci/issues/8446