no need for gem update system anymore
Tests pass on all our tested rubies without it -- it may have been needed for old rubies.

However, it makes our tests on 1.9.3 fail with:

Updating rubygems-update
ERROR:  Error installing rubygems-update:
    rubygems-update requires Ruby version >= 2.3.0.
ERROR:  While executing gem ... (NoMethodError)
    undefined method `version' for nil:NilClass
The command "gem update --system" failed and exited with 1 during

Yes, 1.9.3 is really old, but we don't want to drop it without a major version release, and we'd rather not a major version release, and it's not really hard to keep it supported as we aren't changing ruby-marc much. So we'll fix the test infrastructure this way so tests pass again.