Allow xml parsing to ignore namespace requirements
Currently the REXML parser doesn't require the XML namespace, because REXML.
Many other libraries in other langugages (perl, PHP) will do the same.

This PR adds a new optional keyword argument to `MARC::XMLReader.new`,
`ignore_namespace:`, which allows the XML pull-parsers to parse a
marc-xml file even in the absense of a correct namespace.

The default remains to not parse such files.