Incorporate code from MARC::FastXMLWriter as UnsafeXMLWriter
Makes changes to bring in `MARC::FastXMLWriter` as
`MARC::UnsafeXMLWriter` (to avoid clashes with the
gem). The fast writer just concatenates strings and is
about 15 times faster than the REXML-based version, but
obviously does no validation. In the form of the gem,
this is used "in the wild" across lots of contexts and
I've never heard of anyone having problems, so while the
"unsafe" might be overstating things, people should use it
on purpose and with that knowledge.

The plan is to deprecate the MARC::FastXMLWriter
gem and encourage people to just use MARC::UnsafeXMLWriter
once this is in a release.

  * Minor refactoring to make things easier / share code
  * Add `MARC::Record.to_xml_string` with an optional
    `fast_but_unsafe:` keyword arg indicating use of
    the unsafe writer.