Allow readers to take a custom Record class
Most uses of a MARC record require extraction of things
like OCLC numbers, local identifiers, etc. The ruby-marc
reader classes hard-code the use of MARC::Record, which
means folks who want to do that either need to inline
their code, wrap a MARC::Record object, or call
the reader.

This code allows the binary reader and the non-REXML
XML readers to take a `record_class` option. The
required interface is for the passed class is:

* `#append_control_field(tag, value)`
* `#append_data_field(tag, ind1, ind2, subfields)` where `subfields` is an array of [code, value] pairs
* `#leader` and `#leader=`
* `#new_from_hash(marc_hash)` which is used in the XML reader to avoid using specific field classes

All the rest of the code has been modified to use just those methods when building up a record.

The easiest way to take advantage of this would be to just subclass MARC::Record, e.g.

```
class MyRecord < MARC::Record
  def isbns
    m = /[\d\-Xx]{10,15}/
    fields('020').map{|f| f['a'].scan(m).first.gsub('-', '').upcase}
  end
end

r = MARC::Reader.new("test/batch.dat", record_class: MyRecord).first
r.isbns #=> ["0471383147"]

```